Distribute dbghelp.dll and symsrv.dll with ETW2CTF.
The versions of dbghelp.dll and symsrv.dll distributed with xperf are needed to communicate with a symbol server. There is no guarantee that these libraries are available in 32 bits on the clients that run ETW2CTF, so we distribute them with the converter.

R=etienneb@chromium.org

Review URL: https://codereview.appspot.com/19670043