Use dissectors before other mecanisms to decode the payload of an event.
When a dissector is able to decode the payload of an event, it should be used instead of TDH.

This change allows the Chrome Dissector to be used even if the Chrome MOF is installed. This is important because TDH is not able to decode variable size arrays on some platforms. Also, a dissector can be faster than TDH.

R=etienneb@chromium.org, fw.centrale@gmail.com

Review URL: https://codereview.appspot.com/21250043