Do not crash when an event has no payload.
When an event has no payload, the UserData (payload) field of the
PEVENT_RECORD structure is NULL. This CL removes an assertion that
assumed that this field was never NULL.

The payload is empty for the event from provider MSNT_SystemTrace
(68fdd900-4a3e-11d1-84f4-0000f80464e3) and opcode 0x08 (RDComplete).

I renamed DecodePayloadWithDissectors -> DecodeEventWithDissectors
because a dissector can still be useful when an event has no payload.
It could, for example, associate an event name with the opcode.

TEST=
Generate a kernel trace with "CSWITCH" events and convert it using ETW2CTF.
No assertion should be violated.

R=etienneb@chromium.org, fw.centrale@gmail.com

Review URL: https://codereview.appspot.com/43110043